Add MPL header to pool files and refactor filesystem module
