+++ src/dbus_api/filesystem/fetch_properties_2_0/methods.rs
@@ -5,6 +5,7 @@
 use std::collections::HashMap;
 
 use dbus::{
+    self,
     arg::{Array, RefArg, Variant},
     tree::{MTFn, MethodInfo, MethodResult},
     Message,
@@ -13,7 +14,7 @@ use itertools::Itertools;
 
 use crate::dbus_api::{
     consts,
-    pool::shared::pool_operation,
+    filesystem::shared::filesystem_operation,
     types::TData,
     util::{get_next_arg, result_to_tuple},
 };
@@ -30,22 +31,12 @@ fn get_properties_shared(
     let return_value: HashMap<String, (bool, Variant<Box<dyn RefArg>>)> = properties
         .unique()
         .filter_map(|prop| match prop.as_str() {
-            consts::POOL_TOTAL_SIZE_PROP => Some((
+            consts::FILESYSTEM_USED_PROP => Some((
                 prop,
-                pool_operation(m.tree, object_path.get_name(), |(_, _, pool)| {
-                    Ok((u128::from(*pool.total_physical_size())
-                        * devicemapper::SECTOR_SIZE as u128)
-                        .to_string())
-                }),
-            )),
-            consts::POOL_TOTAL_USED_PROP => Some((
-                prop,
-                pool_operation(m.tree, object_path.get_name(), |(_, _, pool)| {
-                    pool.total_physical_used()
+                filesystem_operation(m.tree, object_path.get_name(), |(_, _, fs)| {
+                    fs.used()
+                        .map(|u| (*u).to_string())
                         .map_err(|e| e.to_string())
-                        .map(|size| {
-                            (u128::from(*size) * devicemapper::SECTOR_SIZE as u128).to_string()
-                        })
                 }),
             )),
             _ => None,
@@ -59,7 +50,7 @@ fn get_properties_shared(
 pub fn get_all_properties(m: &MethodInfo<MTFn<TData>, TData>) -> MethodResult {
     get_properties_shared(
         m,
-        &mut vec![consts::POOL_TOTAL_SIZE_PROP, consts::POOL_TOTAL_USED_PROP]
+        &mut vec![consts::FILESYSTEM_USED_PROP]
             .into_iter()
             .map(|s| s.to_string()),
     )
