+++ src/dbus_api/pool/pool_2_0/api.rs
@@ -1,3 +1,7 @@
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this
+// file, You can obtain one at http://mozilla.org/MPL/2.0/.
+
 use dbus::tree::{Access, EmitsChangedSignal, Factory, MTFn, Method, Property};
 
 use crate::dbus_api::{
